+++ src/autoanki/Dictionary/CC-CEDictionary.py
@@ -1,8 +1,0 @@
-# from Dictionary import Dictionary
-#
-# class CC_CEDictionary(Dictionary):
-#
-#     def __init__(self):
-#         pass
-#
-#