+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Bugfix
 - Fixed imports
 - Added missing sql files to package
+
 ## v1.0.5 (17/04/2023)
 
 ### Feature
@@ -19,3 +20,12 @@
 ## v1.0.0 (17/04/2023)
 
 - First release of `autoanki`
+
+## v1.1.0 (04/05/2024)
+
+- Fixed bugs
+- Removed legacy files
+- Changed dictionary to use local file, rather than internet lookups
+- Formatted logging output
+- Updated README.md
+- Cleaned input to remove punctuation